Format operation + convert lumens
